// apps/web/src/features/props/schemas/prop.schema.ts
import z from 'zod';

export const propFormSchema = z.object({
  name: z.string().trim().min(1, 'Name is required').max(100, 'Name must be 100 characters or less'),
  description: z.string().trim().max(500, 'Description must be 500 characters or less'),
  entityValues: z.array(z.string()),
});

export type PropFormData = z.infer<typeof propFormSchema>;